refactor: language classic game setting
