+++ src/components/game/game.js
@@ -1,5 +1,6 @@
 import { gameState } from '../../../lib/state/state.js';
 import { sendEmailCode } from '../../pages/game.js';
+import applyLanguageClassicSetting from '../language/applyLanguageClassicSetting.js';
 
 export function setupGameSettingModal(page) {
   let gameSettingModal = new bootstrap.Modal(
@@ -30,7 +31,7 @@ export function setupGameSettingModal(page) {
   // 게임 박스 클릭 이벤트
   gameBox.addEventListener('click', function () {
     if (gameState.getState().currentGameStatus === 'idle') {
-      updateGameSettingModalMultilingualContent();
+      applyLanguageClassicSetting()
       gameSettingModal.show();
     }
   });
@@ -38,32 +39,6 @@ export function setupGameSettingModal(page) {
   const sendEmailButton = page.querySelector('#send-email-code-button');
   if (sendEmailButton) {
     sendEmailButton.addEventListener('click', sendEmailCode);
-  }
-}
-
-
-
-export function updateGameSettingModalMultilingualContent() {
-  const elementsToUpdate = {
-    'gameSettingModalLabel': i18next.t('gameSettingModalLabel'),
-    'player2-label': i18next.t('player2-label'),
-    'player-name': i18next.t('player-name'), // placeholder의 경우 추가 처리 필요
-    'mode': i18next.t('mode'),
-    'normal-label': i18next.t('normal-label'),
-    'speed-label': i18next.t('speed-label'),
-    'object-label': i18next.t('object-label'),
-    'startGameButton': i18next.t('startGameButton')
-  };
-
-  for (const [id, text] of Object.entries(elementsToUpdate)) {
-    const element = document.getElementById(id);
-    if (element) {
-      if (id === 'player-name') {
-        element.placeholder = text; // placeholder의 경우 별도 처리
-      } else {
-        element.innerHTML = text;
-      }
-    }
   }
 }
